Add test for issues #11312 and #11325
- https://github.com/nim-lang/Nim/issues/11312
- https://github.com/nim-lang/Nim/issues/11325

+++ nimdoc/tester.nim
@@ -43,21 +43,34 @@ proc testNimDoc(prjDir, docsDir: string; switches: NimSwitches; fixup = false) =
     else:
       echo "SUCCESS: files identical: ", produced
 
-  if failures == 0:
+  if failures == 0 and ((prjDir / docsDir) != prjDir):
     removeDir(prjDir / docsDir)
 
-# Test "nim doc --project .."
+# Test "nim doc --project --outdir:.. --index:on .."
 let
   test1PrjName = "testproject"
   test1Dir = baseDir / test1PrjName
   test1DocsDir = "htmldocs"
   test1Switches = NimSwitches(doc: @["--project",
-                                     "--index:on",
                                      "--outdir:$1/$2" % [test1Dir, test1DocsDir],
+                                     "--index:on",
                                      "$1/$2.nim" % [test1Dir, test1PrjName]],
                               buildIndex: @["--out:$1/$2/theindex.html" % [test1Dir, test1DocsDir],
                                             "$1/$2" % [test1Dir, test1DocsDir]])
 testNimDoc(test1Dir, test1DocsDir, test1Switches, defined(fixup))
+
+# Test "nim doc --out:.. --index:on .."
+let
+  test2PrjDir = "test_out_index_dot_html"
+  test2PrjName = "foo"
+  test2Dir = baseDir / test2PrjDir
+  test2DocsDir = "htmldocs"
+  test2Switches = NimSwitches(doc: @["--out:$1/$2/index.html" % [test2Dir, test2DocsDir],
+                                     "--index:on",
+                                     "$1/$2.nim" % [test2Dir, test2PrjName]],
+                              buildIndex: @["--out:$1/$2/theindex.html" % [test2Dir, test2DocsDir],
+                                            "$1/$2" % [test2Dir, test2DocsDir]])
+testNimDoc(test2Dir, test2DocsDir, test2Switches, defined(fixup))
 
 # Check for failures
 if failures > 0: quit($failures & " failures occurred.")
